requirements.txt
mysql-connector==2.1.6
pandas==0.24.1
Pyrogram==0.11.0
schedule==0.6.0
tgcrypto
